Minor switch changes
Switches now only created if they don't exist so init no longer requires
vmm.ko to be unloaded. Multiple init runs will just do nothing.

Switches have been updated to add a dash to the end of the interface
description. This stops the possibility of the wrong switch being used
if one switch has a name that is a substring of another - eg "int" &
"internet".