Add utctime="no" to Windows guest template
Windows guests (or at least all of them that I have run under bhyve) appear to want local time and not UTC time for their clocks.  As this was the default before 1.2, nothing was needed, but now that the default has changed this is necessary to get the correct time at boot.