Allow manual configuration of boot disk
Allow the user to manually configure a boot disk by specifying
a bootdisk variable with a path as the value.  In particular,
this can be used to work around a limitation that prevents
booting from a virtio-scsi disk in Legacy/BIOS mode with bhyveload.
CTL disks, which are used for virtio-scsi, do not permit ordinary
"read" requests, so bhyveload cannot read from them.  Use this
bootdisk variable to point bhyveload directly to the zvol that
backs the CTL disk used for booting.