vnc: Generate a random password if none specified.
Currently, the VNC server is started without a password if
none was specified in the guest's configuration file (this
is normally done with the "vnc_password" parameter).

This commit changes this behavior to automatically generate
a random password if none is configured. This can be disabled
by specifying a configuration parameter.

There does not appear to be a way to get the password back
to the user via stdout. Since the password already appears
in the guest's log file, I felt that was a sensible place to
store it for now.

In addition to improving the default security of a guest,
this change also allows macOS' builtin VNC client to connect.
The macOS VNC client requires servers to configure a password
(although, it fails to make that obvious).